Improve detection of invalid PS files
A document scanned without errors with no pages and no format is likely
to be an invalid file, or not a PostScript file at all. Since this
cannot be detected by the scanner, in this particular case, we try to
render the document to set SPECTRE_STATUS_LOAD_ERROR in case it fails to
render. Fixes bug #19042.