Use spectre_gs_process() instead of spectre_gs_send_page() in pdf exporter
When rendering with the display device we use a new gs instance for
every page and we have to process prolog, setup and trailer for every
page. PDF exporter is different, we use a single gs instance created in
exporter_pdf_begin() and destroyed in exporter_pdf_end(). Now we use
spectre_gs_process() to send prolog and setup in begin() and trailer in
end() so that we just send the page contents in do_page().
Fixes bug #26592.